pyproject.toml
[project]
name = "rss-adapters"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "python-dateutil>=2.9.0.post0",
    "xmltodict>=1.0.2",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
